chore: update definitions
