+++ manifests/crd.yaml
@@ -862,7 +862,7 @@ spec:
     listKind: CloudStackHostChaosList
     plural: cloudstackhostchaos
     shortNames:
-    - csvm
+    - csh
     singular: cloudstackhostchaos
   scope: Namespaced
   versions:
